+++ test/org/basex/test/storage/UpdateTestTags.java
@@ -22,8 +22,8 @@ public final class UpdateTestTags extends UpdateTest {
     insertTag(3, 0, JUNIT, Data.ELEM);
     assertEquals(size + 1, data.meta.size);
     assertEquals(3, data.parent(4, Data.ELEM));
-    assertEquals((int) Data.ATTR, data.kind(9));
-    assertEquals((int) Data.ELEM, data.kind(5));
+    assertEquals(Data.ATTR, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(5));
     assertEquals(1, data.parent(5, Data.ELEM));
     assertEquals(5, data.parent(6, Data.ELEM));
     assertEquals(nextid + 1, data.meta.lastid);
@@ -31,7 +31,7 @@ public final class UpdateTestTags extends UpdateTest {
     reload();
     assertEquals(size + 1, data.meta.size);
     assertEquals(3, data.parent(4, Data.ELEM));
-    assertEquals((int) Data.ATTR, data.kind(9));
+    assertEquals(Data.ATTR, data.kind(9));
     assertEquals(1, data.parent(5, Data.ELEM));
     assertEquals(5, data.parent(6, Data.ELEM));
     assertEquals(nextid + 1, data.meta.lastid);
@@ -48,8 +48,8 @@ public final class UpdateTestTags extends UpdateTest {
     insertTag(3, 1, JUNIT, Data.ELEM);
     assertEquals(size + 1, data.meta.size);
     assertEquals(3, data.parent(4, Data.ELEM));
-    assertEquals((int) Data.ATTR, data.kind(9));
-    assertEquals((int) Data.ELEM, data.kind(5));
+    assertEquals(Data.ATTR, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(5));
     assertEquals(1, data.parent(5, Data.ELEM));
     assertEquals(5, data.parent(6, Data.ELEM));
     assertEquals(nextid + 1, data.meta.lastid);
@@ -57,7 +57,7 @@ public final class UpdateTestTags extends UpdateTest {
     reload();
     assertEquals(size + 1, data.meta.size);
     assertEquals(3, data.parent(4, Data.ELEM));
-    assertEquals((int) Data.ATTR, data.kind(9));
+    assertEquals(Data.ATTR, data.kind(9));
     assertEquals(1, data.parent(5, Data.ELEM));
     assertEquals(5, data.parent(6, Data.ELEM));
     assertEquals(nextid + 1, data.meta.lastid);
@@ -74,8 +74,8 @@ public final class UpdateTestTags extends UpdateTest {
     insertTag(3, 2, JUNIT, Data.ELEM);
     assertEquals(size + 1, data.meta.size);
     assertEquals(3, data.parent(4, Data.ELEM));
-    assertEquals((int) Data.ATTR, data.kind(9));
-    assertEquals((int) Data.ELEM, data.kind(5));
+    assertEquals(Data.ATTR, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(5));
     assertEquals(1, data.parent(5, Data.ELEM));
     assertEquals(5, data.parent(6, Data.ELEM));
     assertEquals(nextid + 1, data.meta.lastid);
@@ -83,7 +83,7 @@ public final class UpdateTestTags extends UpdateTest {
     reload();
     assertEquals(size + 1, data.meta.size);
     assertEquals(3, data.parent(4, Data.ELEM));
-    assertEquals((int) Data.ATTR, data.kind(9));
+    assertEquals(Data.ATTR, data.kind(9));
     assertEquals(1, data.parent(5, Data.ELEM));
     assertEquals(5, data.parent(6, Data.ELEM));
     assertEquals(nextid + 1, data.meta.lastid);
@@ -99,7 +99,7 @@ public final class UpdateTestTags extends UpdateTest {
     final long nextid = data.meta.lastid;
     insertTag(6, 1, JUNIT, Data.ELEM);
     assertEquals(size + 1, data.meta.size);
-    assertEquals((int) Data.ELEM, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(9));
     assertEquals(6, data.parent(9, Data.ELEM));
     assertByteArraysEqual(JUNIT, data.tag(9));
     assertEquals(6, data.parent(10, Data.ELEM));
@@ -107,7 +107,7 @@ public final class UpdateTestTags extends UpdateTest {
     assertEquals(nextid + 1, data.meta.lastid);
     reload();
     assertEquals(size + 1, data.meta.size);
-    assertEquals((int) Data.ELEM, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(9));
     assertEquals(6, data.parent(9, Data.ELEM));
     assertByteArraysEqual(JUNIT, data.tag(9));
     assertEquals(6, data.parent(10, Data.ELEM));
@@ -124,7 +124,7 @@ public final class UpdateTestTags extends UpdateTest {
     final long nextid = data.meta.lastid;
     insertTag(6, 2, JUNIT, Data.ELEM);
     assertEquals(size + 1, data.meta.size);
-    assertEquals((int) Data.ELEM, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(9));
     assertByteArraysEqual(JUNIT, data.tag(11));
     assertEquals(6, data.parent(11, Data.ELEM));
     assertEquals(6, data.parent(9, Data.ELEM));
@@ -133,7 +133,7 @@ public final class UpdateTestTags extends UpdateTest {
 
     reload();
     assertEquals(size + 1, data.meta.size);
-    assertEquals((int) Data.ELEM, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(9));
     assertByteArraysEqual(JUNIT, data.tag(11));
     assertEquals(6, data.parent(11, Data.ELEM));
     assertEquals(6, data.parent(9, Data.ELEM));
@@ -150,7 +150,7 @@ public final class UpdateTestTags extends UpdateTest {
     final long nextid = data.meta.lastid;
     insertTag(6, 0, JUNIT, Data.ELEM);
     assertEquals(size + 1, data.meta.size);
-    assertEquals((int) Data.ELEM, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(9));
     assertByteArraysEqual(JUNIT, data.tag(11));
     assertEquals(6, data.parent(11, Data.ELEM));
     assertEquals(6, data.parent(9, Data.ELEM));
@@ -158,7 +158,7 @@ public final class UpdateTestTags extends UpdateTest {
     assertEquals(nextid + 1, data.meta.lastid);
     reload();
     assertEquals(size + 1, data.meta.size);
-    assertEquals((int) Data.ELEM, data.kind(9));
+    assertEquals(Data.ELEM, data.kind(9));
     assertByteArraysEqual(JUNIT, data.tag(11));
     assertEquals(6, data.parent(11, Data.ELEM));
     assertEquals(6, data.parent(9, Data.ELEM));
@@ -173,10 +173,10 @@ public final class UpdateTestTags extends UpdateTest {
   public void testUpdateTagName() {
     final Data data = CONTEXT.data();
     data.update(6, token("JUnit"));
-    assertEquals((int) Data.ELEM, data.kind(6));
+    assertEquals(Data.ELEM, data.kind(6));
     assertByteArraysEqual(token("JUnit"), data.tag(6));
     reload();
-    assertEquals((int) Data.ELEM, data.kind(6));
+    assertEquals(Data.ELEM, data.kind(6));
     assertByteArraysEqual(token("JUnit"), data.tag(6));
   }
